feat(show): execute cmd as sql query & show group's creator

+++ server/src/main/kotlin/dao/Dao.kt
@@ -29,5 +29,5 @@ interface Dao<T>{
     /**
      * Get entity by its id
      */
-    fun getById(id: Int): T?
+    fun getById(id: Long): T?
 }
